Week_4_RAG/01_Building_An_LLM_Application/02_Building_Agents/requirements.txt
llama-index-llms-groq
python-dotenv
llama-index
llama-index-tools-yahoo-finance
llama-index-llms-gemini
llama-index-tools-tavily-research